libs/platform/config/pyproject.toml
[project]
name = "casting-config"
version = "0.1.0"
description = "Simple app-level .env loader for Casting development"
requires-python = ">=3.13"
dependencies = [
  "python-dotenv>=1.0.1",
  "pydantic>=2.7",
  "pydantic-settings>=2.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "casting.platform.config"
namespace = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
